Rename spoilers to sync
