+++ .run/Run Spoilers.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run Spoilers" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run -p spoilers --features local-dev" />
+    <option name="command" value="run -p sync --features local-dev" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs />
     <option name="emulateTerminal" value="true" />
