+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
     "packages/contracts",
     "packages/fuzzy",
     "packages/named_lock",
-    "packages/spoilers",
+    "packages/sync",
 ]
 resolver = "2"
