+++ packages/sync/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "spoilers"
+name = "sync"
 version = "0.1.0"
 edition = "2024"
