+++ packages/sync/src/adapters/services/scryfall/data/card.rs
