+++ packages/sync/src/adapters/services/scryfall/data/set.rs
