+++ packages/sync/src/adapters/services/scryfall/utils/image.rs
