+++ packages/sync/src/adapters/services/scryfall/utils/uuid.rs
