+++ packages/sync/src/domain/utils/emoji.rs
