+++ packages/sync/src/domain/utils/images.rs
