+++ packages/sync/src/domain/utils/svg.rs
