+++ packages/sync/src/ports/emoji.rs
