+++ packages/sync/src/ports/image_store.rs
